requirements.txt
fastapi
uvicorn
python-dotenv
moviepy
SpeechRecognition
requests